+++ src/main/java/me/anedhel/lotr/block/ModStoneBlocks.java
@@ -1999,6 +1999,199 @@ public class ModStoneBlocks {
 			.wall(CRACKED_OVERGROWN_GOLD_FANCY_BLUESLATE_BRICKS_WALL)
 			.group("stone").unlockCriterionName("has_cracked_gold_fancy_bricks").build();
 
+    public static final Block CHALK = registerBlock("chalk",
+            new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+    public static final Block CHALK_STAIRS = registerBlock("chalk_stairs",
+            new StairsBlock(CHALK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+    public static final Block CHALK_SLAB = registerBlock("chalk_slab",
+            new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+    public static final Block CHALK_WALL = registerBlock("chalk_wall",
+            new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+    public static final Block CHALK_BUTTON = registerBlock("chalk_button",
+            new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+    public static final Block CHALK_PRESSURE_PLATE = registerBlock("chalk_pressure_plate",
+            new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+                    FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+    public static final BlockFamily CHALK_FAMILY = BlockFamilies.register(CHALK)
+            .stairs(CHALK_STAIRS)
+            .slab(CHALK_SLAB)
+            .wall(CHALK_WALL)
+            .button(CHALK_BUTTON)
+            .pressurePlate(CHALK_PRESSURE_PLATE)
+            .group("stone").unlockCriterionName("has_stone").build();
+
+	public static final Block MOSSY_CHALK = registerBlock("mossy_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block MOSSY_CHALK_STAIRS = registerBlock("mossy_chalk_stairs",
+			new StairsBlock(MOSSY_CHALK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block MOSSY_CHALK_SLAB = registerBlock("mossy_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block MOSSY_CHALK_WALL = registerBlock("mossy_chalk_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final Block MOSSY_CHALK_BUTTON = registerBlock("mossy_chalk_button",
+			new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+	public static final Block MOSSY_CHALK_PRESSURE_PLATE = registerBlock("mossy_chalk_pressure_plate",
+			new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+					FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+	public static final BlockFamily MOSSY_CHALK_FAMILY = BlockFamilies.register(MOSSY_CHALK)
+			.stairs(MOSSY_CHALK_STAIRS)
+			.slab(MOSSY_CHALK_SLAB)
+			.wall(MOSSY_CHALK_WALL)
+			.button(MOSSY_CHALK_BUTTON)
+			.pressurePlate(MOSSY_CHALK_PRESSURE_PLATE)
+			.group("stone").unlockCriterionName("has_stone").build();
+
+	public static final Block OVERGROWN_CHALK = registerBlock("overgrown_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block OVERGROWN_CHALK_STAIRS = registerBlock("overgrown_chalk_stairs",
+			new StairsBlock(OVERGROWN_CHALK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block OVERGROWN_CHALK_SLAB = registerBlock("overgrown_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block OVERGROWN_CHALK_WALL = registerBlock("overgrown_chalk_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final Block OVERGROWN_CHALK_BUTTON = registerBlock("overgrown_chalk_button",
+			new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+	public static final Block OVERGROWN_CHALK_PRESSURE_PLATE = registerBlock("overgrown_chalk_pressure_plate",
+			new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+					FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+	public static final BlockFamily OVERGROWN_CHALK_FAMILY = BlockFamilies.register(OVERGROWN_CHALK)
+			.stairs(OVERGROWN_CHALK_STAIRS)
+			.slab(OVERGROWN_CHALK_SLAB)
+			.wall(OVERGROWN_CHALK_WALL)
+			.button(OVERGROWN_CHALK_BUTTON)
+			.pressurePlate(OVERGROWN_CHALK_PRESSURE_PLATE)
+			.group("stone").unlockCriterionName("has_stone").build();
+
+    public static final Block COBBLED_CHALK = registerBlock("cobbled_chalk",
+            new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+    public static final Block COBBLED_CHALK_STAIRS = registerBlock("cobbled_chalk_stairs",
+            new StairsBlock(COBBLED_CHALK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+    public static final Block COBBLED_CHALK_SLAB = registerBlock("cobbled_chalk_slab",
+            new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+    public static final Block COBBLED_CHALK_WALL = registerBlock("cobbled_chalk_wall",
+            new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+    public static final Block COBBLED_CHALK_BUTTON = registerBlock("cobbled_chalk_button",
+            new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+    public static final Block COBBLED_CHALK_PRESSURE_PLATE = registerBlock("cobbled_chalk_pressure_plate",
+            new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+                    FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+    public static final BlockFamily COBBLED_CHALK_FAMILY = BlockFamilies.register(COBBLED_CHALK)
+            .stairs(COBBLED_CHALK_STAIRS)
+            .slab(COBBLED_CHALK_SLAB)
+            .wall(COBBLED_CHALK_WALL)
+            .button(COBBLED_CHALK_BUTTON)
+            .pressurePlate(COBBLED_CHALK_PRESSURE_PLATE)
+            .group("stone").unlockCriterionName("has_cobbled_stone").build();
+
+	public static final Block MOSSY_COBBLED_CHALK = registerBlock("mossy_cobbled_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block MOSSY_COBBLED_CHALK_STAIRS = registerBlock("mossy_cobbled_chalk_stairs",
+			new StairsBlock(MOSSY_COBBLED_CHALK.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block MOSSY_COBBLED_CHALK_SLAB = registerBlock("mossy_cobbled_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block MOSSY_COBBLED_CHALK_WALL = registerBlock("mossy_cobbled_chalk_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final Block MOSSY_COBBLED_CHALK_BUTTON = registerBlock("mossy_cobbled_chalk_button",
+			new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+	public static final Block MOSSY_COBBLED_CHALK_PRESSURE_PLATE = registerBlock("mossy_cobbled_chalk_pressure_plate",
+			new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+					FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+	public static final BlockFamily MOSSY_COBBLED_CHALK_FAMILY = BlockFamilies.register(MOSSY_COBBLED_CHALK)
+			.stairs(MOSSY_COBBLED_CHALK_STAIRS)
+			.slab(MOSSY_COBBLED_CHALK_SLAB)
+			.wall(MOSSY_COBBLED_CHALK_WALL)
+			.button(MOSSY_COBBLED_CHALK_BUTTON)
+			.pressurePlate(MOSSY_COBBLED_CHALK_PRESSURE_PLATE)
+			.group("stone").unlockCriterionName("has_mossy_cobbled_stone").build();
+
+	public static final Block OVERGROWN_COBBLED_CHALK = registerBlock("overgrown_cobbled_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block OVERGROWN_COBBLED_CHALK_STAIRS = registerBlock("overgrown_cobbled_chalk_stairs",
+			new StairsBlock(OVERGROWN_COBBLED_CHALK.getDefaultState(),
+					FabricBlockSettings.copyOf(Blocks.STONE_STAIRS)));
+	public static final Block OVERGROWN_COBBLED_CHALK_SLAB = registerBlock("overgrown_cobbled_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+	public static final Block OVERGROWN_COBBLED_CHALK_WALL = registerBlock("overgrown_cobbled_chalk_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.ANDESITE_WALL)));
+	public static final Block OVERGROWN_COBBLED_CHALK_BUTTON = registerBlock("overgrown_cobbled_chalk_button",
+			new ButtonBlock(FabricBlockSettings.copyOf(Blocks.STONE_BUTTON), BlockSetType.STONE, 10, false));
+	public static final Block OVERGROWN_COBBLED_CHALK_PRESSURE_PLATE = registerBlock(
+			"overgrown_cobbled_chalk_pressure_plate",
+			new PressurePlateBlock(PressurePlateBlock.ActivationRule.EVERYTHING,
+					FabricBlockSettings.copyOf(Blocks.STONE_PRESSURE_PLATE), BlockSetType.STONE));
+	public static final BlockFamily OVERGROWN_COBBLED_CHALK_FAMILY = BlockFamilies.register(OVERGROWN_COBBLED_CHALK)
+			.stairs(OVERGROWN_COBBLED_CHALK_STAIRS)
+			.slab(OVERGROWN_COBBLED_CHALK_SLAB)
+			.wall(OVERGROWN_COBBLED_CHALK_WALL)
+			.button(OVERGROWN_COBBLED_CHALK_BUTTON)
+			.pressurePlate(OVERGROWN_COBBLED_CHALK_PRESSURE_PLATE)
+			.group("stone").unlockCriterionName("has_overgrown_cobbled_stone").build();
+
+    public static final Block SMOOTH_CHALK = registerBlock("smooth_chalk",
+            new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+    public static final Block SMOOTH_CHALK_SLAB = registerBlock("smooth_chalk_slab",
+            new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+	public static final Block CRACKED_SMOOTH_CHALK = registerBlock("cracked_smooth_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+	public static final Block CRACKED_SMOOTH_CHALK_SLAB = registerBlock("cracked_smooth_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+
+	public static final Block MOSSY_SMOOTH_CHALK = registerBlock("mossy_smooth_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+	public static final Block MOSSY_SMOOTH_CHALK_SLAB = registerBlock("mossy_smooth_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+	public static final Block CRACKED_MOSSY_SMOOTH_CHALK = registerBlock("cracked_mossy_smooth_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+	public static final Block CRACKED_MOSSY_SMOOTH_CHALK_SLAB = registerBlock(
+			"cracked_mossy_smooth_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+
+	public static final Block OVERGROWN_SMOOTH_CHALK = registerBlock("overgrown_smooth_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+	public static final Block OVERGROWN_SMOOTH_CHALK_SLAB = registerBlock("overgrown_smooth_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+	public static final Block CRACKED_OVERGROWN_SMOOTH_CHALK = registerBlock("cracked_overgrown_smooth_chalk",
+			new Block(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE)));
+	public static final Block CRACKED_OVERGROWN_SMOOTH_CHALK_SLAB = registerBlock(
+			"cracked_overgrown_smooth_chalk_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.SMOOTH_STONE_SLAB)));
+
+	public static final Block CHALK_BRICKS = registerBlock("chalk_bricks",
+			new Block(FabricBlockSettings.copyOf(Blocks.STONE_BRICKS)));
+	public static final Block CHALK_BRICKS_STAIRS = registerBlock("chalk_bricks_stairs",
+			new StairsBlock(CHALK_BRICKS.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_BRICK_STAIRS)));
+	public static final Block CHALK_BRICKS_SLAB = registerBlock("chalk_bricks_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_BRICK_SLAB)));
+	public static final Block CHALK_BRICKS_WALL = registerBlock("chalk_bricks_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.STONE_BRICK_WALL)));
+	public static final BlockFamily CHALK_BRICK_FAMILY = BlockFamilies.register(CHALK_BRICKS)
+			.stairs(CHALK_BRICKS_STAIRS)
+			.slab(CHALK_BRICKS_SLAB)
+			.wall(CHALK_BRICKS_WALL)
+			.group("stone").unlockCriterionName("has_brick").build();
+	
+	public static final Block CHISELED_CHALK_BRICK = registerBlock("chiseled_chalk_brick",
+			new Block(FabricBlockSettings.copyOf(Blocks.CHISELED_STONE_BRICKS)));
+
+	public static final Block CRACKED_CHALK_BRICKS = registerBlock("cracked_chalk_bricks",
+			new Block(FabricBlockSettings.copyOf(Blocks.CRACKED_STONE_BRICKS)));
+	public static final Block CRACKED_CHALK_BRICKS_STAIRS = registerBlock("cracked_chalk_bricks_stairs",
+			new StairsBlock(CRACKED_CHALK_BRICKS.getDefaultState(), FabricBlockSettings.copyOf(Blocks.STONE_BRICK_STAIRS)));
+	public static final Block CRACKED_CHALK_BRICKS_SLAB = registerBlock("cracked_chalk_bricks_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_BRICK_SLAB)));
+	public static final Block CRACKED_CHALK_BRICKS_WALL = registerBlock("cracked_chalk_bricks_wall",
+			new WallBlock(FabricBlockSettings.copyOf(Blocks.STONE_BRICK_WALL)));
+	public static final BlockFamily CRACKED_CHALK_BRICK_FAMILY = BlockFamilies.register(CRACKED_CHALK_BRICKS)
+			.stairs(CRACKED_CHALK_BRICKS_STAIRS)
+			.slab(CRACKED_CHALK_BRICKS_SLAB)
+			.wall(CRACKED_CHALK_BRICKS_WALL)
+			.group("stone").unlockCriterionName("has_cracked_brick").build();
+
+	public static final Block CHALK_PILLAR = registerBlock("chalk_pillar",
+			new ModPillarBlock(FabricBlockSettings.copyOf(Blocks.STONE)));
+	public static final Block CHALK_PILLAR_SLAB = registerBlock("chalk_pillar_slab",
+			new SlabBlock(FabricBlockSettings.copyOf(Blocks.STONE_SLAB)));
+
     /**
      * This Method is used to register a Block with minecraft and also register the Item for the block
      * @param name the name of the Block